[EI-999] [web] refactor: Update imports to use 'indicator-value' instead of 'indicator' in various files
