+++ apps/web/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-custom-forecast-indicators-value-view-model.hook.ts
@@ -9,7 +9,7 @@ import { useIndicatorBoard } from '../indicator-board/use-indicator-board.hook';
 import {
   LiveIndicatorRequestParams,
   useFetchLiveIndicatorsValueByType,
-} from '@/app/store/querys/numerical-guidance/indicator.query';
+} from '@/app/store/querys/numerical-guidance/indicator-value.query';
 import { getStartDate } from '@/app/utils/date-formatter';
 
 export const useCustomForecastIndicatorsValueViewModel = (indicatorBoardMetadataId?: string) => {
