+++ apps/web/app/mocks/mock-data/indicators-value.mock.ts
@@ -1,4 +1,4 @@
-import { IndicatorValueResponse } from '@/app/store/querys/numerical-guidance/indicator.query';
+import { IndicatorValueResponse } from '@/app/store/querys/numerical-guidance/indicator-value.query';
 
 export const indicatorsValueMockData: IndicatorValueResponse[] = [
   {
