+++ apps/web/app/store/querys/numerical-guidance/indicator-value.query.ts
